media-processor/requirements.txt
python-socketio>=5.8.0
aiohttp>=3.8.4
aiortc>=1.5.0
opencv-python>=4.7.0
numpy>=1.24.0
pyee>=9.0.4
python-dotenv>=1.0.0
colorlog>=6.7.0
openai>=1.0.0
pillow>=9.0.0 